test(docs): split CLI contract verification by section
